+++ tests/TestCase/Controller/PostsCategoriesControllerTest.php
@@ -140,7 +140,7 @@ class PostsCategoriesControllerTest extends IntegrationTestCase
         }
 
         //Sets the cache name
-        $cache = sprintf('category_%s_limit_%s_page_%s', md5($slug), config('default.records'), 1);
+        $cache = sprintf('category_%s_limit_%s_page_%s', md5($slug), getConfig('default.records'), 1);
         list($postsFromCache, $pagingFromCache) = array_values(Cache::readMany(
             [$cache, sprintf('%s_paging', $cache)],
             $this->PostsCategories->cache