+++ tests/bootstrap.php
@@ -24,6 +24,7 @@ use Cake\Cache\Cache;
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
 use Cake\Datasource\ConnectionManager;
+use Cake\Log\Log;
 use Cake\Mailer\Email;
 use Cake\Routing\DispatcherFactory;
 
@@ -119,12 +120,12 @@ Plugin::load('Assets', [
     'path' => VENDOR . 'mirko-pagliai' . DS . 'assets' . DS,
 ]);
 
-Configure::write('MysqlBackup.connection', 'test');
-Configure::write('MysqlBackup.target', TMP . 'backups');
+Configure::write('DatabaseBackup.connection', 'test');
+Configure::write('DatabaseBackup.target', TMP . 'backups');
 
-Plugin::load('MysqlBackup', [
+Plugin::load('DatabaseBackup', [
     'bootstrap' => true,
-    'path' => VENDOR . 'mirko-pagliai' . DS . 'cakephp-mysql-backup' . DS,
+    'path' => VENDOR . 'mirko-pagliai' . DS . 'cakephp-database-backup' . DS,
 ]);
 
 Configure::write('Tokens.usersClassOptions', [
@@ -162,11 +163,21 @@ Plugin::load('MeCms', [
 
 require_once ROOT . 'config' . DS . 'bootstrap_base.php';
 
+//Sets debug log
+Log::config('debug', [
+    'className' => 'File',
+    'path' => LOGS,
+    'levels' => ['notice', 'info', 'debug'],
+    'file' => 'debug',
+]);
+
 DispatcherFactory::add('Routing');
 DispatcherFactory::add('ControllerFactory');
 
 Email::setConfigTransport('debug', ['className' => 'Debug']);
 Email::setConfig('default', ['transport' => 'debug', 'log' => true]);
+
+Configure::write(DATABASE_BACKUP . '.mailSender', getConfig('email.webmaster'));
 
 ini_set('intl.default_locale', 'en_US');